[PRD-5021] Apply code review directions -cache enabled by default -moved cache flag check to DefaultReportOutputHandlerFactory -fixed cache key computation to work with legacy reports
