+++ src/org/pentaho/reporting/platform/plugin/configuration.properties
@@ -25,3 +25,6 @@ org.pentaho.reporting.engine.classic.extensions.datasources.cda.CdaQueryBackend=
 # Tell the engine to use the pentaho provider by default. This will make all uses of the mondrian system use
 # the pentaho cube-file provider.
 org.pentaho.reporting.engine.classic.extensions.datasources.mondrian.CubeFileProvider=org.pentaho.reporting.platform.plugin.connection.PentahoCubeFileProvider
+
+#Cache for pageble HTML
+org.pentaho.reporting.platform.plugin.output.CachePageableHtmlContent=true
